refactor(site): fold `<Badges />` into `Badge/PresetBadges`
Enabled, Premium, and the other labeled badges are thin Badge wrappers, not a second primitive. Move them into Badge/PresetBadges and rename the layout wrapper to BadgeGroup.
