+++ site/src/components/Badge/Badge.stories.tsx
@@ -5,9 +5,9 @@ import {
 	SettingsIcon,
 	TriangleAlertIcon,
 } from "lucide-react";
-import { Badges } from "#/components/Badges/Badges";
 import { Spinner } from "#/components/Spinner/Spinner";
 import { Badge } from "./Badge";
+import { BadgeGroup } from "./PresetBadges";
 
 const meta: Meta<typeof Badge> = {
 	title: "components/Badge",
@@ -18,7 +18,7 @@ type Story = StoryObj<typeof Badge>;
 
 export const Default: Story = {
 	render: () => (
-		<Badges>
+		<BadgeGroup>
 			<Badge size="xs">
 				<DatabaseIcon />
 				Text
@@ -31,13 +31,13 @@ export const Default: Story = {
 				<DatabaseIcon />
 				Text
 			</Badge>
-		</Badges>
+		</BadgeGroup>
 	),
 };
 
 export const Warning: Story = {
 	render: () => (
-		<Badges>
+		<BadgeGroup>
 			<Badge variant="warning" size="xs">
 				Warning
 				<TriangleAlertIcon />
@@ -50,13 +50,13 @@ export const Warning: Story = {
 				<TriangleAlertIcon />
 				Warning
 			</Badge>
-		</Badges>
+		</BadgeGroup>
 	),
 };
 
 export const Destructive: Story = {
 	render: () => (
-		<Badges>
+		<BadgeGroup>
 			<Badge variant="destructive" size="xs">
 				Destructive
 				<TriangleAlertIcon />
@@ -69,13 +69,13 @@ export const Destructive: Story = {
 				<TriangleAlertIcon />
 				Destructive
 			</Badge>
-		</Badges>
+		</BadgeGroup>
 	),
 };
 
 export const Info: Story = {
 	render: () => (
-		<Badges>
+		<BadgeGroup>
 			<Badge variant="info" size="xs">
 				Info
 			</Badge>
@@ -85,13 +85,13 @@ export const Info: Story = {
 			<Badge variant="info" size="md">
 				Info
 			</Badge>
-		</Badges>
+		</BadgeGroup>
 	),
 };
 
 export const Green: Story = {
 	render: () => (
-		<Badges>
+		<BadgeGroup>
 			<Badge variant="green" size="xs">
 				Green
 			</Badge>
@@ -101,13 +101,13 @@ export const Green: Story = {
 			<Badge variant="green" size="md">
 				Green
 			</Badge>
-		</Badges>
+		</BadgeGroup>
 	),
 };
 
 export const Purple: Story = {
 	render: () => (
-		<Badges>
+		<BadgeGroup>
 			<Badge variant="purple" size="xs">
 				Purple
 			</Badge>
@@ -117,13 +117,13 @@ export const Purple: Story = {
 			<Badge variant="purple" size="md">
 				Purple
 			</Badge>
-		</Badges>
+		</BadgeGroup>
 	),
 };
 
 export const Magenta: Story = {
 	render: () => (
-		<Badges>
+		<BadgeGroup>
 			<Badge variant="magenta" size="xs">
 				Magenta
 			</Badge>
@@ -133,7 +133,7 @@ export const Magenta: Story = {
 			<Badge variant="magenta" size="md">
 				Magenta
 			</Badge>
-		</Badges>
+		</BadgeGroup>
 	),
 };
 
@@ -157,7 +157,7 @@ export const MediumWithIcon: Story = {
 
 export const StatusWithIcon: Story = {
 	render: () => (
-		<Badges>
+		<BadgeGroup>
 			<Badge variant="info" size="md" role="status">
 				<Spinner loading />
 				Running
@@ -170,6 +170,6 @@ export const StatusWithIcon: Story = {
 				<TriangleAlertIcon />
 				Failed
 			</Badge>
-		</Badges>
+		</BadgeGroup>
 	),
 };
