+++ site/src/components/Badge/PresetBadges.tsx
@@ -1,6 +1,7 @@
-import { Badge } from "#/components/Badge/Badge";
+import type { ComponentPropsWithRef, FC, PropsWithChildren } from "react";
+import { Badge } from "./Badge";
 
-export const EnabledBadge: React.FC = () => {
+export const EnabledBadge: FC = () => {
 	return (
 		<Badge className="option-enabled" variant="green">
 			Enabled
@@ -8,7 +9,7 @@ export const EnabledBadge: React.FC = () => {
 	);
 };
 
-export const DisabledBadge: React.FC<React.ComponentPropsWithRef<"div">> = ({
+export const DisabledBadge: FC<ComponentPropsWithRef<"div">> = ({
 	...props
 }) => {
 	return (
@@ -18,29 +19,25 @@ export const DisabledBadge: React.FC<React.ComponentPropsWithRef<"div">> = ({
 	);
 };
 
-export const EnterpriseBadge: React.FC = () => {
+export const EnterpriseBadge: FC = () => {
 	return <Badge variant="purple">Enterprise</Badge>;
 };
 
-interface PremiumBadgeProps {
-	children?: React.ReactNode;
-}
-
-export const PremiumBadge: React.FC<PremiumBadgeProps> = ({
+export const PremiumBadge: FC<PropsWithChildren> = ({
 	children = "Premium",
 }) => {
 	return <Badge variant="magenta">{children}</Badge>;
 };
 
-export const AlphaBadge: React.FC = () => {
+export const AlphaBadge: FC = () => {
 	return <Badge variant="purple">Alpha</Badge>;
 };
 
-export const DeprecatedBadge: React.FC = () => {
+export const DeprecatedBadge: FC = () => {
 	return <Badge variant="warning">Deprecated</Badge>;
 };
 
-export const Badges: React.FC<React.PropsWithChildren> = ({ children }) => {
+export const BadgeGroup: FC<PropsWithChildren> = ({ children }) => {
 	return (
 		<div className="flex flex-row items-center gap-2 mb-4">{children}</div>
 	);
