+++ site/src/pages/DeploymentSettingsPage/NetworkSettingsPage/NetworkSettingsPageView.tsx
@@ -1,10 +1,10 @@
 import type { FC } from "react";
 import type { SerpentOption } from "#/api/typesGenerated";
 import {
-	Badges,
+	BadgeGroup,
 	DisabledBadge,
 	EnabledBadge,
-} from "#/components/Badges/Badges";
+} from "#/components/Badge/PresetBadges";
 import {
 	SettingsHeader,
 	SettingsHeaderDescription,
@@ -60,14 +60,14 @@ export const NetworkSettingsPageView: FC<NetworkSettingsPageViewProps> = ({
 				</SettingsHeaderDescription>
 			</SettingsHeader>
 
-			<Badges>
+			<BadgeGroup>
 				{useDeploymentOptions(options, "Wildcard Access URL")[0].value !==
 				"" ? (
 					<EnabledBadge />
 				) : (
 					<DisabledBadge />
 				)}
-			</Badges>
+			</BadgeGroup>
 		</div>
 	</div>
 );
