+++ site/src/pages/DeploymentSettingsPage/Option.tsx
@@ -1,6 +1,6 @@
 import { WrenchIcon } from "lucide-react";
 import type { FC, HTMLAttributes, PropsWithChildren } from "react";
-import { DisabledBadge, EnabledBadge } from "#/components/Badges/Badges";
+import { DisabledBadge, EnabledBadge } from "#/components/Badge/PresetBadges";
 import { cn } from "#/utils/cn";
 
 export const OptionName: FC<PropsWithChildren> = ({ children }) => {
