+++ site/src/pages/DeploymentSettingsPage/SecuritySettingsPage/SecuritySettingsPageView.tsx
@@ -1,10 +1,10 @@
 import type { FC } from "react";
 import type { SerpentOption } from "#/api/typesGenerated";
 import {
-	Badges,
+	BadgeGroup,
 	DisabledBadge,
 	EnabledBadge,
-} from "#/components/Badges/Badges";
+} from "#/components/Badge/PresetBadges";
 import {
 	SettingsHeader,
 	SettingsHeaderDescription,
@@ -68,9 +68,9 @@ export const SecuritySettingsPageView: FC<SecuritySettingsPageViewProps> = ({
 						className="items-center"
 					>
 						Browser-Only Connections{" "}
-						<Badges>
+						<BadgeGroup>
 							{featureBrowserOnlyEnabled ? <EnabledBadge /> : <DisabledBadge />}
-						</Badges>
+						</BadgeGroup>
 					</SettingsHeaderTitle>
 					<SettingsHeaderDescription>
 						Block all workspace access via SSH, port forward, and other
